vein-pay-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
pycryptodome
tensorflow
opencv-python
numpy
python-multipart
